no strict 'syms'; check illegal names
Add a new strict check syms
Disallow illegal unparsable symbol and class names.

Esp. interesting are embedded \0 in classnames, which were until 5.16
silently ignored, and since 5.16 allowed. Since 5.16 names are
internally nul-safe, but roundtrips are not supported, and
such hidden payloads are useless for perl, are hard to detect and
may lead to security problems.