run-time part of sized arrays. my @a[20];
sized arrays cannot grow beyond the declared size and cannot be magic.
The declared size is always equal to the actual size, the array is pre-filled
with undef.
thus sized arrays are faster to access at run-time than aelemfast (constant indices),
but this name is already taken.

add 2 new ops: aelemsize_const for constant indices, similar to aelemfast_lex
and aelemsize for variable indices.
constant indices are already compile-time checked, there is no need to
bounds-check and check magic at run-time.

access to non-existing indices result in a die with PL_no_aelem.
these can be yet unassigned indices or indices beyond the declared size.

negative indices for constants are resolved at compile-time, as the
actual size is already known then, so we can omit the check for negative indices
at run-time.
Instead of being pre-filled with &PL_sv_undef we could imagine dying on access
on an unassigned index.

Without AELEMSIZE_RT_NEGATIVE the declared size is always equal to the actual size.
To allow the actual size (AvFILL) deviate from the declared size (AvMAX)
but this causes a slower run-time access for constant indices. (percent?)

Bump Opcodes to 1.28 for the 2 new ops in tag :base_core