fix perl #125840: $x=*0; *x=$x
In this case the lhs of the glob references the rhs, and when
clearing the lhs before the copy of the GP values, the rhs was deleted.
Protect from that.

Furthermore, the result needs to be assigned a fresh GP to fix
the case with -e'*x=$x=*a;$a=$x'

fixup glob_assign_glob